Extract common test setup for Genesis
- New module `Test.Ouroboros.Consensus.Genesis.Setup`
- Move `genChains` from its own module to `Genesis.Setup`
- Move `TestSetup`-related code from `LongRangeAttack` to `Genesis.Setup`
