+++ ouroboros-consensus/test/infra-test/Test/Ouroboros/Consensus/Genesis/Tests/LongRangeAttack.hs
@@ -1,20 +1,14 @@
 {-# LANGUAGE BlockArguments      #-}
 {-# LANGUAGE DerivingStrategies  #-}
-{-# LANGUAGE NamedFieldPuns      #-}
 {-# LANGUAGE RecordWildCards     #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 
 module Test.Ouroboros.Consensus.Genesis.Tests.LongRangeAttack (tests) where
 
-import           Control.Monad.Class.MonadTime (MonadTime)
-import           Control.Monad.Class.MonadTimer.SI (MonadTimer)
 import           Control.Monad.IOSim (runSimOrThrow)
-import           Control.Tracer (traceWith)
 import qualified Data.Map.Strict as Map
 import           Ouroboros.Consensus.Block.Abstract hiding (Header)
 import           Ouroboros.Consensus.Config
-import           Ouroboros.Consensus.Util.Condense
-import           Ouroboros.Consensus.Util.IOLike
 import           Ouroboros.Network.AnchoredFragment (headAnchor)
 import qualified Ouroboros.Network.AnchoredFragment as AF
 import qualified Ouroboros.Network.AnchoredFragment.Extras as AF
@@ -26,10 +20,8 @@ import           Test.Ouroboros.Consensus.ChainGenerator.Tests.Adversarial hidin
 import qualified Test.Ouroboros.Consensus.ChainGenerator.Tests.BlockTree as BT
 import           Test.Ouroboros.Consensus.ChainGenerator.Tests.BlockTree
                      (BlockTreeBranch (btbFull))
-import           Test.Ouroboros.Consensus.ChainGenerator.Tests.GenChain
-                     (genChains)
 import           Test.Ouroboros.Consensus.ChainGenerator.Tests.PointSchedule
-import           Test.Ouroboros.Consensus.PeerSimulator.Run
+import           Test.Ouroboros.Consensus.Genesis.Setup
 import qualified Test.QuickCheck as QC
 import           Test.QuickCheck
 import           Test.QuickCheck.Random (QCGen)
@@ -37,28 +29,29 @@ import           Test.Tasty
 import           Test.Tasty.QuickCheck
 import           Test.Util.Orphans.IOLike ()
 import           Test.Util.TestBlock hiding (blockTree)
-import           Test.Util.Tracer (recordingTracerTVar)
 
 tests :: TestTree
-tests = testProperty "long range attack" prop_syncGenesis
+tests = testProperty "long range attack" prop_longRangeAttack
 
-prop_syncGenesis :: SomeTestAdversarial -> QCGen -> QC.Property
-prop_syncGenesis (SomeTestAdversarial _ _ params) seed =
+prop_longRangeAttack :: SomeTestAdversarial -> QCGen -> QC.Property
+prop_longRangeAttack (SomeTestAdversarial _ _ params) seed =
   withMaxSuccess 10 $
-    runSimOrThrow $ runTest params (exampleTestSetup params seed)
+    runSimOrThrow $ runTest setup $ \fragment ->
+      pure
+        $ classify genesisAfterIntersection "Long range attack"
+        $ classify genesisAcrossIntersection "Genesis potential"
+        $ genesisAfterIntersection ==> not $ isHonestTestFragH fragment
 
-newtype GenesisWindow = GenesisWindow { getGenesisWindow :: SlotNo }
-  deriving stock (Show)
+  where
+    setup@TestSetup{..} = exampleTestSetup params seed
 
-data TestSetup = TestSetup {
-    secParam                  :: SecurityParam
-  , genesisWindow             :: GenesisWindow
-  , schedule                  :: PointSchedule
-  , blockTree                 :: BT.BlockTree TestBlock
-  , genesisAcrossIntersection :: Bool
-  , genesisAfterIntersection  :: Bool
-  }
-  deriving stock (Show)
+    isHonestTestFragH :: TestFragH -> Bool
+    isHonestTestFragH frag = case headAnchor frag of
+        AF.AnchorGenesis   -> True
+        AF.Anchor _ hash _ -> isHonestTestHeaderHash hash
+
+    isHonestTestHeaderHash :: HeaderHash TestBlock -> Bool
+    isHonestTestHeaderHash = all (0 ==) . unTestHash
 
 exampleTestSetup ::
   TestAdversarial base hon ->
@@ -83,48 +76,10 @@ exampleTestSetup params seed =
     fragLenA = AF.slotLength badChain
     advLenAfterIntersection = AF.length badChainSuffix
 
+    honestAsc = testAscH params
+
     blockTree = genChains (testAscA params) (testRecipeA params) (testRecipeA' params) seed
     goodChain = BT.btTrunk blockTree
 
     BT.BlockTreeBranch { btbSuffix = badChainSuffix, btbFull = badChain } =
       head $ BT.btBranches blockTree
-
-runTest ::
-  (IOLike m, MonadTime m, MonadTimer m) =>
-  TestAdversarial base hon ->
-  TestSetup ->
-  m Property
-runTest TestAdversarial{testAscH, testAscA} TestSetup{..} = do
-    (tracer, getTrace) <- recordingTracerTVar
-
-    traceWith tracer "Active slot coefficient: "
-    traceWith tracer $ "  honest: "++ show testAscH
-    traceWith tracer $ "  adversary: "++ show testAscA
-
-    mapM_ (traceWith tracer) $ BT.prettyPrint blockTree
-
-    let advPeer = PeerId "adversary"
-    let peers = [HonestPeer, advPeer]
-
-    result <- runPointSchedule secParam testAscH schedule tracer blockTree peers
-    trace <- unlines <$> getTrace
-
-    pure
-      $ classify genesisAfterIntersection "Long range attack"
-      $ classify genesisAcrossIntersection "Genesis potential"
-      $ counterexample trace
-      $ case result of
-          Left exn ->
-            counterexample ("exception: " <> show exn) False
-          Right frag ->
-            counterexample ("result: " <> condense frag)
-            $ genesisAfterIntersection ==> not $ isHonestTestFragH frag
-
-  where
-    isHonestTestFragH :: TestFragH -> Bool
-    isHonestTestFragH frag = case headAnchor frag of
-        AF.AnchorGenesis   -> True
-        AF.Anchor _ hash _ -> isHonestTestHeaderHash hash
-
-    isHonestTestHeaderHash :: HeaderHash TestBlock -> Bool
-    isHonestTestHeaderHash = all (0 ==) . unTestHash
